[fix] Update ace interface file names miracl_workflow_ace_interface.py used to contain an older interface. Update this file to take the contents of ace_test_interface_abstract.py
